+++ intellistop/libs/__init__.py
@@ -9,7 +9,9 @@ from .beta import get_beta, get_daily_gains
 from .alpha import get_alpha
 from .k_ratio import get_k_ratio
 from .variances import calculate_variances, calculate_time_series_variances
-from .vq import run_vq_calculation, find_latest_max, get_stop_loss_from_value
+from .volatility_factor import (
+    run_vq_calculation, find_latest_max, get_stop_loss_from_value, generate_stop_loss_data_set
+)
 from .filters import (
     windowed_filter, subtraction_filter, simple_moving_average_filter,
     exponential_moving_average_filter, weighted_moving_average_filter, smart_moving_average,
