+++ intellistop/libs/lib_types.py
@@ -96,6 +96,8 @@ class VQStopsResultType:
     stop_loss: VQStopLossResultType
     current_max: float
     fund_name: str
+    stop_loss_data_set: list
+    event_log: list
 
     def __init__(self):
         self.fund_name = ""
@@ -104,6 +106,8 @@ class VQStopsResultType:
         self.derived = VQStopLossRawResultType()
         self.vq = VQStopLossResultType()
         self.stop_loss = VQStopLossResultType()
+        self.stop_loss_data_set = []
+        self.event_log = []
 
 
 class SmartMovingAvgType:
@@ -115,6 +119,22 @@ class SmartMovingAvgType:
         self.data_set = []
         self.short_slope = []
         self.long_slope = []
+
+
+class StopLossEventType(Enum):
+    stop = 'stop'
+    minimum = 'minimum'
+    activate = 'activate'
+
+class StopLossEventLogType:
+    index: int
+    event: StopLossEventType
+    price: float
+
+    def __init__(self):
+        self.index = 0
+        self.event = StopLossEventType.stop
+        self.price = 0.0
 
 
 class BetaPropertyEnum(Enum):
